+++ src/ft_memcpy.c
@@ -6,13 +6,13 @@
 /*   By: mclaudel <mclaudel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/08 13:51:25 by mclaudel          #+#    #+#             */
-/*   Updated: 2019/10/08 14:09:03 by mclaudel         ###   ########.fr       */
+/*   Updated: 2019/10/08 15:58:03 by mclaudel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <libft.h>
 
-void	*ft_memcpy(void *dst, const void *src, unsigned int n)
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char *dest;
 	char *source;
