+++ src/ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: mclaudel <mclaudel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/07 14:12:55 by mclaudel          #+#    #+#             */
-/*   Updated: 2019/10/08 13:34:17 by mclaudel         ###   ########.fr       */
+/*   Updated: 2019/10/08 15:58:19 by mclaudel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <libft.h>
 
-void	*ft_memset(void *b, int c, unsigned int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char *src;
